Nuke warnings 	* src/liburbi/uobject/urbi/ubinary.hh, 	* src/liburbi/uobject/uvalue-common.cc: 	Use size_t for sizes, not int.
	* src/examples/monitor.cc,
	* src/examples/urbiballtrackinghead.cc,
	* src/examples/urbiping.cc,
	* src/examples/urbitalkie.cc,
	* src/liburbi/uconversion.cc,
	* src/liburbi/usyncclient.cc:
	Various fixes to address compiler warnings.

git-svn-id: file:///srv/svn/liburbi-cpp/trunk@355 b8db33d8-8104-0410-b827-8403f563b8eb

+++ sdk-remote/src/examples/monitor.cc
@@ -108,13 +108,16 @@ Monitor::processMessages()
 		if ((*it)->window == event.xexpose.window)
 		  {
 		    (*it)->put();
-		    printf("repainting %d\n", event.xexpose.window);
+		    // FIXME: What can be done here instead of these
+		    // two casts?
+		    printf("repainting %d\n",
+			   static_cast<int>(event.xexpose.window));
 		    found = true;
 		    break;
 		  }
 	      if (!found)
 		printf("error: expose event for unknown window %d\n",
-		       event.xexpose.window);
+		       static_cast<int>(event.xexpose.window));
 	    }
 	}
     }
@@ -171,7 +174,7 @@ Monitor::Monitor(int _w, int _h, const char * name, bool _fastMode)
 
   createWindow(name ? name : "XImage - XShm optimized");
   pthread_mutex_unlock(&lock);
-  printf("Monitor created window %d\n", window);
+  printf("Monitor created window %d\n", static_cast<int>(window));
 }
 
 Monitor::~Monitor()
